# api/openapi.yaml
openapi: 3.1.0
info:
  title: Tabverse sync API
  version: 0.1.0
  description: |
    Contract for `tabversed`, the Tabverse sync server.

    ## Authentication
    Two credentials, and they are not interchangeable.

    * **The extension** sends `Authorization: Bearer <device token>` on
      `/api/v1/sync`, `/api/v1/entities/*` and `/api/v1/search`. It pairs
      (`POST /api/v1/auth/pair`) and never signs in: no password, no session,
      no cookie (ADR 0002).
    * **The console** (`/api/v1/admin/*`, `/api/v1/console/*`) uses the session
      cookie set by the auth library, plus an `X-XSRF-Token` header on every
      state changing request. A device token is refused there (ADR 0013).

    ## Console sessions
    A console sign-in lasts `TABVERSED_SESSION_TTL` (default `24h`) and then
    ends: the server refuses the token's expiry rather than letting the auth
    library renew it, and clears the cookie. `0` disables the bound.
    `POST /api/v1/console/revoke-sessions` ends every session of the calling
    account by moving its revocation cut-off, and the operator's
    `PUT /api/v1/admin/users/{user_id}/revoke-sessions` does the same for
    somebody else. Neither touches the extension's device tokens, which are a
    different credential with no expiry (ADR 0021).

    ## Accounts
    Accounts are always on; there is no master credential. People register with
    a single-use email link (or a configured social provider) and manage their
...

    ## Revisions
...

...

...

...
    ```
...
servers:
...
paths:
...
  /api/v1/auth/bootstrap:
    post:
      tags: [auth]
      summary: Create the first account of a deployment
      description: |
        Only succeeds while the server has zero users; afterwards it returns
        409 and the deployment must be paired with an invite code.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, default: default }
      responses:
        '201':
          description: account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceCredentials' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/auth/pair:
    post:
      tags: [auth]
      summary: Redeem a single use pairing code for a device token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invite_code]
              properties:
                invite_code: { type: string, examples: ['BSKT-VT6V-7SX3-XVCQ'] }
                device_name: { type: string }
      responses:
        '201':
          description: device registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceCredentials' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/invites:
    post:
      tags: [auth]
      summary: Mint a pairing code for another device of this account
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ttl_seconds:
                  { type: integer, minimum: 60, maximum: 86400, default: 900 }
      responses:
        '201':
          description: code created (single use)
          content:
            application/json:
              schema:
                type: object
                required: [code, expires_at]
                properties:
                  code: { type: string, examples: ['BSKT-VT6V-7SX3-XVCQ'] }
                  expires_at: { type: integer, description: unix ms }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/console/pair:
    post:
      tags: [console]
      summary: Mint a device and its token for the signed-in account
      description: |
        The official-server pairing wizard (adr/0020). The extension opens the
        console in a window with its own id and a nonce in the URL; the person
        signs in there, names the device and approves it; this endpoint creates
        the device and its token **for the session's account**, and the page
        hands the token to the extension over `chrome.runtime.sendMessage` (the
        extension's `externally_connectable` channel), never rendering it.

        Security: the console session cookie and its XSRF header, not a device
        token. There is **no account parameter**: the account is the session's,
        so a page that could name one would be the invite endpoint with extra
        steps. An assumed identity (impersonation) is refused, as on every other
        mutating console route - a token issued under one would outlive it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_name]
              properties:
                device_name:
                  { type: string, examples: ['Chrome on Windows'] }
                extension_id:
                  type: string
                  description: >-
                    the extension that asked, recorded on the device row so the
                    account's device list says which browser paired it
      responses:
        '201':
          description: device + token, returned once
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceCredentials' }
        '401':
          description: no session, or the session does not resolve to an account
          content:
            application/json:
              schema: { type: object }
        '403':
          description: read only - the caller is impersonating an account
          content:
            application/json:
              schema: { type: object }
  /api/v1/console/revoke-sessions:
    post:
      tags: [console]
      summary: End every console session on this account
      description: |
        "Sign out everywhere" (adr/0021): moves the account's revocation
        cut-off, so every session it ever issued - including this browser's,
        which is why the response also clears the cookie - has to sign in
        again. Audited as `sessions_revoked`.

        **The extension's device tokens are not touched.** They are a different
        credential with no expiry, and revoking them would silently stop
        somebody's devices syncing; unpair them from the devices endpoints
        instead. A cut-off is compared at second granularity, so a session
        minted in the same second survives it.
      responses:
        '204': { description: every session of this account is now refused }
        '401':
          description: no session, or the session does not resolve to an account
          content:
            application/json:
              schema: { type: object }
        '403':
          description: read only - the caller is impersonating an account
          content:
            application/json:
              schema: { type: object }
  /api/v1/sync:
    get:
      tags: [sync]
      summary: Delta download of records changed since a revision
      parameters:
        - {
            name: since,
            in: query,
            schema: { type: integer, minimum: 0, default: 0 },
          }
        - { name: limit, in: query, schema: { type: integer, minimum: 1 } }
      responses:
        '200':
          description: records ordered by revision
          content:
            application/json:
              schema:
                type: object
                required: [records, next_rev, has_more, server_rev]
                properties:
                  records:
                    type: array
                    items: { $ref: '#/components/schemas/Record' }
                  next_rev:
                    { type: integer, description: pass as `since` next time }
                  has_more: { type: boolean }
                  server_rev: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [sync]
      summary: Batch upload of records (LWW per record)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [records]
              properties:
                records:
                  type: array
                  maxItems: 500
                  items: { $ref: '#/components/schemas/RecordInput' }
      responses:
        '200':
          description: per record outcome
          content:
            application/json:
              schema:
                type: object
                required: [results, server_rev]
                properties:
                  results:
                    type: array
                    items: { $ref: '#/components/schemas/RecordResult' }
                  server_rev: { type: integer }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { $ref: '#/components/responses/TooLarge' }
  /api/v1/entities/{entity}/{id}:
    parameters:
      - { $ref: '#/components/parameters/entity' }
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [sync]
      summary: Fetch one record (tombstones included)
      responses:
        '200':
          description: the record
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Record' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [sync]
      summary: Soft delete (creates a tombstone propagated by delta sync)
      responses:
        '200': { $ref: '#/components/responses/SyncResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/search:
    get:
      tags: [search]
      summary: Full text search over this account's records
      description: |
        Every term is matched, not just one of them, and the last term is
        matched as a prefix so the extension can search while the user types.

        Hits come back ranked, and each one names the **tabverse** it belongs
        to (`tabspace_id`), because that is the unit the extension lists: a
        match on one tab, note or closed tab means "this tabverse matched".
        The client drops the tabverses it has not downloaded, so a hit for a
        tabverse this device has never synced is expected and harmless.

        `allnote` / `alltodo` / `allbookmark` are never indexed (they are id
        lists) and `session` only exists for pre-ADR-0006 clients.
      parameters:
        - {
            name: q,
            in: query,
            required: true,
            schema: { type: string },
            description: 'free text; last term is matched as a prefix',
          }
        - {
            name: entity,
            in: query,
            schema: { $ref: '#/components/schemas/Entity' },
          }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 200 },
          }
      responses:
        '200':
          description: ranked hits
          content:
            application/json:
              schema:
                type: object
                required: [query, hits]
                properties:
                  query: { type: string }
                  hits:
                    type: array
                    items: { $ref: '#/components/schemas/SearchHit' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/sync/stream:
    get:
      tags: [sync]
      summary: Realtime change notifications (WebSocket upgrade)
      description: |
        Authenticate with the `Authorization` header or `?access_token=`.
        See the top level description for the frame format.
      parameters:
        - { name: access_token, in: query, schema: { type: string } }
      responses:
        '101': { description: switching protocols }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/admin/totals:
    get:
      tags: [admin]
      summary: Deployment wide counters (the console header)
      responses:
        '200':
          description: totals
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Totals' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AdminDisabled' }
  /api/v1/admin/users:
    get:
      tags: [admin]
      summary: "The operator's directory: every account, with its counters"
      description: |
        Operator-only. A person signing in sees their own account and nothing
        else, which is why this is a forbidden view for them. The address and
        the role are here because without them a row is an opaque id (ADR
        0014).
      parameters:
        - {
            name: q,
            in: query,
            schema: { type: string },
            description: substring of the name or the email address
          }
      responses:
        '200':
          description: accounts
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items: { $ref: '#/components/schemas/UserSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AdminDisabled' }
  # There is no POST /api/v1/admin/users: registration is the only way an
  # account is created (ADR 0014), so there is exactly one path that can prove
  # an address.
  /api/v1/admin/users/{user_id}:
    parameters: [{ $ref: '#/components/parameters/userId' }]
    get:
      tags: [admin]
      summary: One account with its stats, devices and tokens
      responses:
        '200':
          description: account detail (user, stats, devices, tokens)
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [admin]
      summary: Rename an account (its data is untouched)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '204': { description: renamed }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [admin]
      summary: Delete an account and everything it stored
      description: |
        Records, devices, tokens, invites and the account's search index rows
        all go. There is no trash; the console asks for a typed confirmation.
      responses:
        '204': { description: deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/users/{user_id}/invites:
    parameters: [{ $ref: '#/components/parameters/userId' }]
    post:
      tags: [admin]
      summary: Mint a pairing code for this account
      description: |
        Lets an operator add a device to an account they did not create. The
        code is single use and short lived; `POST /api/v1/auth/pair` exchanges
        it for a device token.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ttl_seconds: { type: integer, minimum: 60, maximum: 86400 }
      responses:
        '201':
          description: the code, shown once
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: { type: string }
                  code: { type: string }
                  expires_at: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/users/{user_id}/devices/{device_id}:
    parameters:
      - { $ref: '#/components/parameters/userId' }
      - { name: device_id, in: path, required: true, schema: { type: string } }
    delete:
      tags: [admin]
      summary: Revoke every live token of one device
      responses:
        '204': { description: revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/users/{user_id}/tokens/{hash}:
    parameters:
      - { $ref: '#/components/parameters/userId' }
      - name: hash
        in: path
        required: true
        schema: { type: string }
        description: the stored hash from the account's token listing
    delete:
      tags: [admin]
      summary: Revoke one device token
      description: |
        Tokens are stored only as a SHA-256 hash, so the console can show a
        token by fingerprint and revoke it by hash, but never display it again.
      responses:
        '204': { description: revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/users/{user_id}/revoke-sessions:
    parameters:
      - { $ref: '#/components/parameters/userId' }
    put:
      tags: [admin]
      summary: End every console session on this account (operator's move)
      description: |
        The operator's "sign out everywhere", for the case where the account to
        fix is not the one at the keyboard (adr/0021). Moves the account's
        revocation cut-off and audits `sessions_revoked` against both the actor
        and the account. Device tokens are untouched - see
        `POST /api/v1/console/revoke-sessions`.
      responses:
        '204': { description: every session of this account is now refused }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/users/{user_id}/tokens/{hash}/archive:
    parameters:
      - { $ref: '#/components/parameters/userId' }
      - { name: hash, in: path, required: true, schema: { type: string } }
    put:
      tags: [admin]
      summary: Archive a revoked token
      description: |
        Hides the token from the console's default lists. Requires it to be
        revoked first: archiving is tidying, revoking is how access is cut.
        The token stays revoked, and it stays in use or not exactly as it was
        (ADR 0011).
      responses:
        '200': { description: archived, or a no-op if it already was }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: the token is not revoked yet
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [admin]
      summary: Unarchive a token (it stays revoked)
      responses:
        '200': { description: unarchived }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/users/{user_id}/devices/{device_id}/archive:
    parameters:
      - { $ref: '#/components/parameters/userId' }
      - { name: device_id, in: path, required: true, schema: { type: string } }
    put:
      tags: [admin]
      summary: Archive a device (revoke and archive its tokens, then archive it)
      description: |
        One deterministic teardown: every token the device holds is revoked,
        those tokens are archived, and the device is archived, in a single
        transaction. Immediate - there is no inactivity window to wait out and
        no need to revoke first. Hides the rows from the console's default
        lists; they stay stored and their records keep syncing (ADR 0011).
      responses:
        '200': { description: archived, or a no-op if it already was }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [admin]
      summary: Unarchive a device (its tokens stay revoked)
      responses:
        '200': { description: unarchived }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/users/{user_id}/devices/{device_id}/records/archive:
    parameters:
      - { $ref: '#/components/parameters/userId' }
      - { name: device_id, in: path, required: true, schema: { type: string } }
    put:
      tags: [admin]
      summary: Archive the records this device last wrote
      description: |
        Requires the device to have no usable token left, which archiving the
        device satisfies. The records stay stored and keep syncing to the
        user's own devices, and the extension's search is not filtered - only
        this console's views are. A record edited again un-archives itself.
        Tombstones are left alone.
      responses:
        '200':
          description: how many records were archived
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArchiveOutcome' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: still usable (revoke it first, or archive the device)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [admin]
      summary: Unarchive the records this device wrote
      responses:
        '200':
          description: how many records were restored
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArchiveOutcome' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: |
        A **device token** - the extension's credential, issued by pairing. It is
        not accepted on the console routes, which use a session cookie instead.
...
  schemas:
...
    DeviceCredentials:
      type: object
      required: [user_id, device_id, token, server_rev]
      properties:
        user_id: { type: string }
        device_id: { type: string }
        token: { type: string, description: store this, shown only once }
        server_rev: { type: integer }
        issued_at: { type: integer }
...
  responses:
...
    Unauthorized:
      description: missing or invalid token / invite code
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: account already bootstrapped
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
security:
  - bearerToken: []